requirements.txt
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
graphviz==0.8.4
idna==2.8
joblib==0.14.0
kiwisolver==1.1.0
matplotlib==3.2.0
mxnet-cu92==1.6.0b20191101
mxnet-cu92mkl==1.6.0b20191031
numpy==1.14.6
packaging==19.2
pandas==0.25.3
pkg-resources==0.0.0
pyparsing==2.4.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
scikit-learn==0.22.2.post1
scipy==1.3.1
six==1.13.0
urllib3==1.25.6
